Have number of followers

+++ app/Controllers/Http/FollowsController.ts
@@ -1,24 +1,31 @@
 import { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import APIException from 'App/Exceptions/APIException'
 import Follow from 'App/Models/Follow'
+import User from 'App/Models/User'
 
 export default class FollowController {
   public async follow({ params, response }: HttpContextContract) {
     const followerId = params.followerId
     const followingId = params.followingId
 
-    // Vérifier si la relation de suivi existe déjà
+    const targetUser = await User.findBy('id', followingId)
+
+    if (!targetUser) {
+      throw new APIException("L'utilisateur à suivre est inconnu")
+    }
+
     const existingFollow = await Follow.query()
       .where('follower_id', followerId)
       .where('following_id', followingId)
       .first()
 
     if (!existingFollow) {
-      // Créer la relation de suivi
       await Follow.create({
         followerId,
         followingId,
       })
+
+      targetUser.followers += 1
 
       return response.noContent()
     }
@@ -30,11 +37,18 @@ export default class FollowController {
     const followerId = params.followerId
     const followingId = params.followingId
 
-    // Supprimer la relation de suivi
+    const targetUser = await User.findBy('id', followingId)
+
+    if (!targetUser) {
+      throw new APIException("L'utilisateur à suivre est inconnu")
+    }
+
     await Follow.query()
       .where('follower_id', followerId)
       .where('following_id', followingId)
       .delete()
+
+    targetUser.followers -= 1
 
     return response.noContent()
   }
